Hide java project setting files by default.
