+++ package.json
@@ -89,6 +89,12 @@
           "description": "Specifies the severity of the message when the classpath is incomplete for a Java file",
           "scope": "window"
         },
+        "java.configuration.excludeProjectSettingFiles": {
+          "type": "boolean",
+          "default": true,
+          "description": "Exclude the extension generated project setting files from file explorer.",
+          "scope": "window"
+        },
         "java.configuration.updateBuildConfiguration": {
           "type": [
             "string"
