+++ src/utils.ts
@@ -1,0 +1,7 @@
+'use strict';
+
+import { workspace, WorkspaceConfiguration } from 'vscode';
+
+export function getJavaConfiguration(): WorkspaceConfiguration {
+	return workspace.getConfiguration('java');
+}
